requirements.txt
fastapi
uvicorn
pyspark
onetl[spark]
etl-entities
sqlalchemy
psycopg2-binary
pandas
python-dotenv
kaggle
pyyaml
